Renamed gem from number_to_words_ru to numbers_and_words.
