--- lib/number_to_words_ru/array_additions.rb
@@ -1,2 +1,0 @@
-require 'number_to_words_ru/array_additions/helpers'
-require 'number_to_words_ru/array_additions/validations'
