--- lib/number_to_words_ru/strategies.rb
@@ -1,3 +1,0 @@
-require 'number_to_words_ru/strategies/base'
-require 'number_to_words_ru/strategies/ru'
-require 'number_to_words_ru/strategies/en'
