--- lib/number_to_words_ru/translations_helpers.rb
@@ -1,3 +1,0 @@
-require 'number_to_words_ru/translations_helpers/base'
-require 'number_to_words_ru/translations_helpers/ru'
-require 'number_to_words_ru/translations_helpers/en'
