+++ lib/numbers_and_words/array_additions.rb
@@ -1,0 +1,2 @@
+require 'numbers_and_words/array_additions/helpers'
+require 'numbers_and_words/array_additions/validations'
