+++ lib/numbers_and_words/array_additions/helpers.rb
