+++ lib/numbers_and_words/array_additions/validations.rb
