+++ lib/numbers_and_words/core_ext/array.rb
