+++ lib/numbers_and_words/core_ext/integer.rb
