+++ lib/numbers_and_words/figures_array.rb
