+++ lib/numbers_and_words/strategies.rb
@@ -1,0 +1,3 @@
+require 'numbers_and_words/strategies/base'
+require 'numbers_and_words/strategies/ru'
+require 'numbers_and_words/strategies/en'
