+++ lib/numbers_and_words/strategies/base.rb
