+++ lib/numbers_and_words/strategies/en.rb
