+++ lib/numbers_and_words/strategies/ru.rb
