+++ lib/numbers_and_words/translations_helpers.rb
@@ -1,0 +1,3 @@
+require 'numbers_and_words/translations_helpers/base'
+require 'numbers_and_words/translations_helpers/ru'
+require 'numbers_and_words/translations_helpers/en'
