+++ lib/numbers_and_words/translations_helpers/en.rb
