+++ lib/numbers_and_words/translations_helpers/ru.rb
