+++ spec/numbers_and_words/array/ru_spec.rb
