+++ spec/numbers_and_words/integer/ru_spec.rb
