+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 $TESTING=true
 $:.unshift File.join(File.dirname(__FILE__), '..', 'lib')
 
-require 'number_to_words_ru'
+require 'numbers_and_words'
 
 RSpec.configure do
   include TranslationsHelpers::Base
